refactor(codegen): extract examples into plugin folder
