+++ scripts/codegen/src/core/registry.ts
@@ -1,5 +1,6 @@
 import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
+import { examplesPlugin } from "../plugins/examples/index.ts";
 import { motionPlugin } from "../plugins/motion/index.ts";
 import { tokensPlugin } from "../plugins/tokens/index.ts";
 import { variantsPlugin } from "../plugins/variants/index.ts";
@@ -16,4 +17,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   variantsPlugin,
   dependenciesPlugin,
   coveragePlugin,
+  examplesPlugin,
 ];
