+++ scripts/codegen/src/plugins/examples/schema.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const exampleEntry = z.strictObject({
+  id: z.string().min(1),
+  props: z.record(z.string(), z.unknown()).optional(),
+});
