+++ scripts/codegen/src/semantic-checks.test.ts
@@ -9,7 +9,6 @@ import {
   checkElementByProp,
   checkEvents,
   checkEventsRuntimeSupport,
-  checkExamplesPresent,
   checkFormControl,
   checkImperativeProps,
   checkPolymorphicAtomic,
@@ -800,26 +799,6 @@ describe("checkPolymorphicAtomic", () => {
       },
     });
     expect(checkPolymorphicAtomic(spec)).toEqual([]);
-  });
-});
-
-describe("checkExamplesPresent", () => {
-  test("flags a spec with no `examples:` block", () => {
-    const issues = checkExamplesPresent(makeButton());
-    expect(issues).toHaveLength(1);
-    expect(issues[0]?.path).toBe("examples");
-  });
-
-  test("flags a spec with an empty examples list", () => {
-    const issues = checkExamplesPresent(makeButton({ examples: [] }));
-    expect(issues).toHaveLength(1);
-  });
-
-  test("passes when at least one example is declared", () => {
-    const issues = checkExamplesPresent(
-      makeButton({ examples: [{ id: "default", props: { intent: "primary" } }] }),
-    );
-    expect(issues).toEqual([]);
   });
 });
